requirements.txt
discord.py
feedparser
translate
python-dotenv